doc: Change path of Adapter interface
Channel will not be exposed through D-Bus, a command line parameter
should be added to allow the users to change the default page and
channel. In order to reduce complexity, physical adapter will not
be directly represented through D-Bus APIs. It is being considered,
only one "wpan0" and one "lowpan0" interface will be available for
each physical adapter.